Fixed no public methods Warning label and Editable Label margins
